% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneWayTable1.R
\name{onewaytable1}
\alias{onewaytable1}
\title{One way table function}
\usage{
onewaytable1(InputTable, alpha, plottilte)
}
\arguments{
\item{InputTable}{data for table to be analyze}

\item{alpha}{to calculate confidence interval}

\item{plottilte}{title for plot}
}
\value{
returns a list the point stimates, the confidence intervals, and the difference confidence intervals for each cell
}
\description{
One way table function
}
