% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoWayTable.R
\name{twowaytable}
\alias{twowaytable}
\title{Two way table}
\usage{
twowaytable(InputTable, alpha)
}
\arguments{
\item{InputTable}{pre-processed table to be analyze. use xtab}

\item{alpha}{alpha for the confidence interval calculations}
}
\value{
a list with confidence intervals, chi tests, and many other uselful data.
}
\description{
Two way table
}
